refactor fetch command handling
